Wired up Tests for Cash deposit and linked to Front End.  Missing return link to resolve

+++ PortfolioManager2/PortfolioManager.Repository/Entities/Account.cs
@@ -12,7 +12,7 @@ namespace PortfolioManager.Repository.Entities
 
         public string Name { get; set; }
 
-        public int Cash { get; set; }
+        public decimal Cash { get; set; }
         public decimal Valuation { get; set; }
 
         public int PortfolioId { get; set; }
